Normal map support
Added support of normal map in the G-Buffer, we now use a proper PBR material workflow. Most normal maps will require their green channel (Y) to be flipped since most of them are created with the D3D coordinates system in mind, which mean PBR textures for UE4 for example will most likely need to use this fix.
Something still seems to be wrong since some maps appears to be mirrored when lit (i.e. a lit area that seems to be darkened) but well...