Envmap switching
Added the ability to switch between differents envmaps in the GUI.
Currently buttons pointing to hardcoded envmaps paths. Here we use the same method to create texture to update the envmap since they may not be using the same size/ratio/color depth... which make the usage of glTexSubImage2D not so easy.
We could render to texture using dedicated FBO but hey, since it is not a critical need...